requirements.txt
absl-py==2.3.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
attr==0.3.1
attrs==25.3.0
azure-core==1.34.0
azure-storage-blob==12.25.1
black==25.1.0
bleach==5.0.1
blinker==1.9.0
blis==1.3.0
boto==2.49.0
boto3==1.38.24
botocore==1.38.24
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==45.0.3
cymem==2.0.11
datamodel-code-generator==0.26.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
Django==5.1.9
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
expiringdict==1.2.2
Faker==37.3.0
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
ftfy==6.3.1
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.4.0
importlib_metadata==8.6.1
inflect==5.6.2
inflection==0.5.1
isodate==0.7.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsf==0.11.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
label-studio==1.18.0
label-studio-sdk==1.0.12
langcodes==3.5.0
language_data==1.3.0
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
lxml==5.4.0
lxml_html_clean==0.4.2
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.41.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
openai==1.82.0
opentelemetry-api==1.33.1
ordered-set==4.0.2
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-mock==1.12.1
rich==14.0.0
rpds-py==0.25.1
rq==1.16.2
rsa==4.9.1
rstr==3.2.2
rules==3.4
s3transfer==0.13.0
scikit-learn==1.6.1
scipy==1.15.3
semver==3.0.4
sentry-sdk==2.29.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.6
threadpoolctl==3.6.0
tldextract==5.3.0
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.40.0
wrapt==1.17.2
xmljson==0.2.1
zipp==3.22.0